backend/pyproject.toml
[tool.poetry]
name = "grocery-aid"
version = "0.1.0"
description = "App aiding grocery shopping in Finland"
authors = ["Jaakko Moisio <jaakko@moisio.fi>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
gql = {extras = ["aiohttp"], version = "^3.1.0"}
SQLAlchemy = {extras = ["postgesql_asyncpg"], version = "^1.4.32"}
asyncpg = "^0.25.0"
SQLAlchemy-Utils = "^0.38.2"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.1.0"
mypy = "^0.941"
pylint = "^2.12.2"
sqlalchemy-stubs = "^0.4"
factory-boy = "^3.2.1"
pytest-asyncio = "^0.18.2"
aiosqlite = "^0.17.0"

[tool.mypy]
plugins = "sqlmypy"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
filterwarnings = "ignore::sqlalchemy.exc.SAWarning"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
